+++ src/Interface/AppInterface/ProjectDockWidget.h
@@ -80,7 +80,7 @@ private:
 
   // HANDLEFILESIZECHANGED:
   // this function handles the signal that the file size has changed and sets the ui to reflect it
-  static void HandleFileSizeChanged( qpointer_type qpointer, double file_size );
+  static void HandleFileSizeChanged( qpointer_type qpointer, long long file_size );
 
 private:
   // SET_AUTO_SAVE_LABEL:
@@ -97,7 +97,7 @@ private:
 
   // SET_FILE_SIZE_LABEL:
   // function that actually sets the label to reflect the file size change
-  void set_file_size_label( double file_size );
+  void set_file_size_label( long long file_size );
   
 private Q_SLOTS:
   // ENABLE_LOAD_DELETE_AND_EXPORT_BUTTONS:
